requirements.txt
alembic==1.11.1
anyio==3.7.0
async-exit-stack==1.0.1
async-generator==1.10
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
